+++ about-us.html
@@ -64,5 +64,8 @@
         <footer>
                 <p class="right-justification"><a href="./about-us.html">&copy MIT Grads</a></p>
         </footer>
+        <div id="special">
+            <a href="/?theme=dirty">X</a>
+        </div>
     </body>
 </html>
