+++ index.html
@@ -24,7 +24,7 @@
             </div>
         </header>
         <main>
-            <div id="splash-image">
+            <div id="staff-place-holder">
                 <figure>
                     <img src="assets/splash.gif" alt="image of musical intervals">
                     <p>Click to Begin</p>
@@ -36,6 +36,7 @@
                 <p>You may listen to an interval again by pressing the "Play Interval" button.</p>
                 <p>When you've made your choice click the "Submit" button to go on to a new interval.</p>
                 <p>At the end of the quiz your results will be displayed.</p>
+                </p>
                 <button id="show-settings-button">Show Optional Quiz Settings</button>
                 <form id="landing-page-form">
                     <div id="settings-form" class="hidden">
