+++ src/data/interval-class.js
@@ -12,6 +12,7 @@ export class IntervalClass {
             this.scale = chromaticScale;
         }
     }
+    
     getFirstNote() {
         return this.firstNote;
     }
