+++ src/playInterval.js
@@ -1,10 +1,14 @@
-import { playNote } from '../utility-functions/playNote.js';
+import { playNote } from './playNote.js';
+
 
 export function playInterval(firstNote, secondNote, instrument, type = 'melodic', duration) {
 
     if(type === 'melodic') {
+
         playNote(instrument, firstNote, duration);
+
         const pause = (duration * 1000);
+
         setTimeout(() => {
             playNote(instrument, secondNote, duration);
         }, pause);
@@ -12,6 +16,9 @@ export function playInterval(firstNote, secondNote, instrument, type = 'melodic'
     } else if(type === 'harmonic') {
         playNote(instrument, firstNote, duration);
         playNote(instrument, secondNote, duration);
+
     }
+
+
 }
