+++ src/quiz-page/quiz-page.css
@@ -7,7 +7,6 @@ body {
     background-repeat: no-repeat;
     background-size: cover;
 }
-
 main {
     position: relative;
     margin: 0;
@@ -43,7 +42,6 @@ main {
     flex-wrap: wrap;
     justify-content: center;
 }
-
 .choices div {
     font-size: 2rem;
     display: inline-flex;
@@ -60,24 +58,22 @@ main {
     border-radius: 15px;
     cursor: pointer;
 }
-
 .choices div p {
     font-family: 'Oleo Script', sans-serif;
     padding: 0 0 .5em 0;
 }
-
 .choices div:hover {
     transform: translate(-2px, -2px);
     background-color: #F1F1D4;
     box-shadow:2px 2px 5px 1px #ABA8B2;
     color: black;
 }
-
 .choices div.selected {
     background-color: #F1F1D4;
     border: 2px solid#6893E2 ;
     color: black;
 }
+
 
 .answer-container {
     display: flex;
@@ -89,23 +85,19 @@ main {
     align-items: center;
     padding-top: 20px;
 }
-
 .answer-container.correct {
     animation: pulse 400ms;
     animation-iteration-count: infinite;
 }
-
 .answer-container.correct-jarring {
     animation: heartBeat 1000ms;
     animation-timing-function: ease-in;
     animation-iteration-count: infinite;
 }
-
 .answer-container.incorrect {
     animation: jiggle 80ms;
     animation-iteration-count: infinite;
 }
-
 .answer-container.incorrect-jarring {
     animation: shake 200ms;
     animation-timing-function: cubic-bezier(.36,.07,.19,.97);
@@ -121,7 +113,6 @@ main {
         transform: rotate(1deg);
     }
 }
-
 @keyframes shake {
     10%, 90% {
       transform: translate3d(-15px, 11px, 0);
@@ -135,7 +126,7 @@ main {
     40%, 60% {
       transform: translate3d(15px, -11px, 0);
     }
-}
+  }
 
 #instructions-display {
     position: absolute;
@@ -143,7 +134,6 @@ main {
     bottom: 5px;
     cursor: pointer;
 }
-
 .instructions-slider {
     width: 30%;
     min-width: 180px;
@@ -158,7 +148,6 @@ main {
     transition: 450ms;
     transition-timing-function: ease-in-out;
 }
-
 .instructions-slider.displayed {
     top: 0;
 }
