+++ src/utils.js
@@ -1,3 +1,4 @@
+
 export function findById(items, id) {
     for(let i = 0; i < items.length; i++) {
         const item = items[i];
@@ -7,6 +8,8 @@ export function findById(items, id) {
     }
     return null;
 }
+
+
 
 // Fisher-Yates Shuffle. Source: https://javascript.info/task/shuffle
